+++ src/lib/components/Spoiler.react.js
@@ -55,24 +55,6 @@ Spoiler.propTypes = {
     maxHeight: PropTypes.number,
 
     /**
-     * Object that holds the loading state object coming from dash-renderer
-     */
-    loading_state: PropTypes.shape({
-        /**
-         * Determines if the component is loading or not
-         */
-        is_loading: PropTypes.bool,
-        /**
-         * Holds which property is loading
-         */
-        prop_name: PropTypes.string,
-        /**
-         * Holds the name of the component that is loading
-         */
-        component_name: PropTypes.string,
-    }),
-
-    /**
      * Label for open spoiler action
      */
     showLabel: PropTypes.string.isRequired,